% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anacapa_utils.R
\name{categorize_continuous_metadata}
\alias{categorize_continuous_metadata}
\title{Takes a metadata file, and categorizes any continuous variable into "low, med, high"}
\usage{
categorize_continuous_metadata(metadata_file)
}
\arguments{
\item{metadata_file}{a well-formatted metadata file}
}
\value{
a metadata file with categorical column in place of the continuous one
}
\description{
Takes a metadata file, and categorizes any continuous variable into "low, med, high"
}
\author{
Gaurav Kandlikar
}
