% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anacapa_utils.R
\name{categorize_continuous_vector}
\alias{categorize_continuous_vector}
\title{Takes a continuous vector and categorizes the vector into low, medium, and high}
\usage{
categorize_continuous_vector(vec)
}
\arguments{
\item{vec}{a continuous vector to categorize}
}
\value{
a continuous vector
}
\description{
Takes a continuous vector and categorizes the vector into low, medium, and high
}
\author{
Gaurav Kandlikar
}
