% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anacapa_utils.R
\name{convert_anacapa_to_phyloseq}
\alias{convert_anacapa_to_phyloseq}
\title{Takes an site-abundance table from Anacapa, along with a qiime-style mapping file, and returns a phyloseq object}
\usage{
convert_anacapa_to_phyloseq(ana_out, mapping_file)
}
\arguments{
\item{ana_out}{OTU table from Anacapa}

\item{mapping_file}{Qiime-style mapping}
}
\value{
phyloseq class object
}
\description{
Takes an site-abundance table from Anacapa, along with a qiime-style mapping file, and returns a phyloseq object
}
\author{
Gaurav Kandlikar
}
