% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_biom_to_taxon_table.R
\name{convert_biom_to_taxon_table}
\alias{convert_biom_to_taxon_table}
\title{Takes a biom table imported in using phyloseq::import_biom() and converts it
into an Anacapa-formmated taxonomy table}
\usage{
convert_biom_to_taxon_table(physeq_object)
}
\arguments{
\item{physeq_object}{phyloseq object with a OTU table and a Taxon Table embedded}
}
\description{
Takes a biom table imported in using phyloseq::import_biom() and converts it
into an Anacapa-formmated taxonomy table
}
\author{
Gaurav Kandlikar
}
