% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_rarefaction.R
\name{custom_rarefaction}
\alias{custom_rarefaction}
\title{Rarefy a phyloseq object to a custom sample depth and with given number of repliactes}
\usage{
custom_rarefaction(physeq_object, sample_size = 10000, replicates = 10,
  ...)
}
\arguments{
\item{physeq_object}{A phyloseq class object to be rarefied}

\item{sample_size}{Desired depth of rarefaction}

\item{replicates}{Desired number of times to resample from community}

\item{...}{Other options for phyloseq function rarefy_even_depth}
}
\value{
phyloseq class object
}
\description{
Rarefy a phyloseq object to a custom sample depth and with given number of repliactes
}
\examples{
good_taxon_table <- data.frame(sum.taxonomy = c("a;b;c;d;f;u", "p;q;r;s;t;u"),
 site_1 = c(0,1), site_2 = c(10, 20))
good_maps <- data.frame(site = c("site_1", "site_2"),
 season = c("wet", "dry"), host = c("oak", "sage"))
physeq_object <- convert_anacapa_to_phyloseq(good_taxon_table, good_maps)
custom_rarefaction(physeq_object, sample_size = 10, replicates = 1)
}
