% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrare.R
\name{ggrare}
\alias{ggrare}
\title{Make a rarefaction curve using ggplot2}
\usage{
ggrare(physeq, step = 10, label = NULL, color = NULL, plot = TRUE,
  parallel = FALSE, se = TRUE)
}
\arguments{
\item{step}{Step Size for sample size in rarefaction curves}

\item{label}{Default `NULL`. Character string. The name of the variable to map to text labels on the plot. Similar to color option but for plotting text.}

\item{color}{Default `NULL`. Character string. The name of the variable to map to the colors in the plot. This can be a sample variables (among the set returned by `sample_variables(physeq)` ) or taxonomic rank (among the set returned by `rank_names(physeq)`).}

\item{plot}{default `TRUE`. Logical. Should the graph be plotted}

\item{parallel}{default `FALSE`. Logical. Should rarefaction be parallelized}

\item{se}{default `TRUE`.Logical. Should standard errors be calculated.}

\item{color}{default `NULL`.Character string. The name of the variable to map to the colors in the plot. Similar to color option but for plotting text.}
}
\description{
Make a rarefaction curve using ggplot2
}
