% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_adonis.R
\name{pairwise_adonis}
\alias{pairwise_adonis}
\title{Pairwise multilevel comparison using adonis}
\usage{
pairwise_adonis(x, factors, sim_method = "bray",
  p_adjust_m = "bonferroni", reduce = NULL)
}
\arguments{
\item{x}{Data frame (the community table).}

\item{factors}{Vector (a column or vector with the levels to be compared pairwise).}

\item{sim_method}{Similarity method from vegdist, default is 'bray'}

\item{p_adjust_m}{The p.value correction method, one of the methods supported by p.adjust(),
default is 'bonferroni'.}

\item{reduce}{String. Restrict comparison to pairs including these factors. If more than one factor, separate by pipes like:  reduce = 'setosa|versicolor'}
}
\value{
Table with the pairwise factors, F-values, R^2, p.value and adjusted p.value.
}
\description{
This is a wrapper function for multilevel pairwise comparison
using adonis() from package 'vegan'. The function returns adjusted p-values using p.adjust().
}
\examples{
data(iris)
pairwise_adonis(iris[, 1:4], iris$Species)

pairwise_adonis(iris[, 1:4], iris$Species, reduce = 'setosa')

# similarity euclidean from vegdist and holm correction
pairwise_adonis(x = iris[, 1:4], factors = iris$Species,
sim_method = 'euclidian', p_adjust_m = 'holm')

#similarity manhattan from daisy and bonferroni correction
pairwise_adonis(x = iris[, 1:4], factors = iris$Species,
sim_method = 'manhattan', p_adjust_m = 'bonferroni')
}
