% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physeq_to_vegan.R
\name{physeq_to_vegan}
\alias{physeq_to_vegan}
\title{Converts the OTU table in a phyloseq object into a Vegan-style community matrix}
\usage{
physeq_to_vegan(physeq)
}
\arguments{
\item{physeq}{`phyloseq` class object}
}
\value{
Vegan-style community matrix (species in columns; sites in rows)
}
\description{
Converts the OTU table in a phyloseq object into a Vegan-style community matrix
}
\author{
Gaurav Kandlikar
}
