% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_validators.R
\name{scrub_seqNum_column}
\alias{scrub_seqNum_column}
\title{scrub_seqNum_col takes one biom table as its input, and if it include
a column named "xxx_seq_number", it gets rid of that column - it's not of use to us
any longer}
\usage{
scrub_seqNum_column(ana_out)
}
\description{
scrub_seqNum_col takes one biom table as its input, and if it include
a column named "xxx_seq_number", it gets rid of that column - it's not of use to us
any longer
}
\author{
Gaurav Kandlikar
}
