% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_validators.R
\name{scrub_taxon_paths}
\alias{scrub_taxon_paths}
\title{Replace empty calls in Anacapa taxonomy tables with Unknown
(that is what they effectively are to most users)}
\usage{
scrub_taxon_paths(taxon_table)
}
\arguments{
\item{taxon_table}{taxonomy table from Anacapa}
}
\value{
ana_taxon_table with scrubbed 'sum.taxonomy' column
}
\description{
Replace empty calls in Anacapa taxonomy tables with Unknown
(that is what they effectively are to most users)
}
\author{
Gaurav Kandlikar
}
