% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_adonis.R
\name{summary.pwadonis}
\alias{summary.pwadonis}
\title{Summarize the results of pairwise_adonis}
\usage{
\method{summary}{pwadonis}(object, ...)
}
\arguments{
\item{object}{output from pairwise_adonis}

\item{...}{other options for print}
}
\description{
Summarize the results of pairwise_adonis
}
