% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anacapa_utils.R
\name{convert_anacapa_to_phyloseq}
\alias{convert_anacapa_to_phyloseq}
\title{Takes an site-abundance table from Anacapa, along with a qiime-style mapping file, and returns a phyloseq object}
\usage{
convert_anacapa_to_phyloseq(taxon_table, metadata_file)
}
\arguments{
\item{taxon_table}{Taxon table in the anacapa format}

\item{metadata_file}{Metadata file with rows as sites, columns as variables}
}
\value{
phyloseq class object
}
\description{
Takes an site-abundance table from Anacapa, along with a qiime-style mapping file, and returns a phyloseq object
}
\examples{
good_taxon_table <- data.frame(sum.taxonomy = c("a;b;c;d;f;u", "p;q;r;s;t;u"),
site_1 = c(0,1), site_2 = c(10, 20))
good_maps <- data.frame(site = c("site_1", "site_2"), season = c("wet", "dry"),
 host = c("oak", "sage"))
convert_anacapa_to_phyloseq(good_taxon_table, good_maps)
}
\author{
Gaurav Kandlikar
}
