% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anacapa_utils.R
\name{convert_anacapa_to_phyloseq}
\alias{convert_anacapa_to_phyloseq}
\title{Takes an site-abundance table from Anacapa, along with a qiime-style mapping file, and returns a phyloseq object}
\usage{
convert_anacapa_to_phyloseq(taxon_table, metadata_file,
  cols_to_categorize = NULL)
}
\arguments{
\item{taxon_table}{Taxon table in the anacapa format}

\item{metadata_file}{Metadata file with rows as sites, columns as variables}

\item{cols_to_categorize}{a vector containing the names of any columns in metadata_file that should be categorized into "high, medium, low"}
}
\value{
phyloseq class object
}
\description{
Takes an site-abundance table from Anacapa, along with a qiime-style mapping file, and returns a phyloseq object
}
\author{
Gaurav Kandlikar
}
