% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_rarefaction.R
\name{custom_rarefaction}
\alias{custom_rarefaction}
\title{Rarefy a phyloseq object to a custom sample depth and with given number of repliactes}
\usage{
custom_rarefaction(physeq, sample_size = 1e+05, replicates = 10, ...)
}
\arguments{
\item{physeq}{A phyloseq class object to be rarefied}

\item{sample_size}{Desired depth of rarefaction}

\item{replicates}{Desired number of times to resample from community}
}
\value{
phyloseq class object
}
\description{
Rarefy a phyloseq object to a custom sample depth and with given number of repliactes
}
\author{
Gaurav Kandlikar
}
