% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anacapa_utils.R
\name{group_anacapa_by_taxonomy}
\alias{group_anacapa_by_taxonomy}
\title{Takes a site-abundance table from Anacapa, and summarizes to each unique taxon in the sum.taxonomy column}
\usage{
group_anacapa_by_taxonomy(ana_out)
}
\arguments{
\item{ana_out}{OTU table from Anacapa}
}
\description{
Takes a site-abundance table from Anacapa, and summarizes to each unique taxon in the sum.taxonomy column
}
\author{
Gaurav Kandlikar
}
