% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_validators.R
\name{scrub_seqNum_column}
\alias{scrub_seqNum_column}
\title{Remove "xxx_seq_number" column from ana_taxon_table file if it exists
takes one taxon table as its input, and if it include
a column named "xxx_seq_number", it gets rid of that column - it's not of use to us
any longer}
\usage{
scrub_seqNum_column(taxon_table)
}
\arguments{
\item{taxon_table}{OTU table from Anacapa}
}
\value{
ana_taxon_table file, with "xxx_seq_number" column removed (if it existed)
}
\description{
Remove "xxx_seq_number" column from ana_taxon_table file if it exists
takes one taxon table as its input, and if it include
a column named "xxx_seq_number", it gets rid of that column - it's not of use to us
any longer
}
\author{
Gaurav Kandlikar
}
