% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_validators.R
\name{scrub_taxon_paths}
\alias{scrub_taxon_paths}
\title{Replace empty calls in Anacapa taxonomy tables with Unknown
(that is what they effectively are to most users)}
\usage{
scrub_taxon_paths(taxon_table)
}
\arguments{
\item{taxon_table}{taxonomy table from Anacapa}
}
\value{
ana_taxon_table with scrubbed 'sum.taxonomy' column
}
\description{
Replace empty calls in Anacapa taxonomy tables with Unknown
(that is what they effectively are to most users)
}
\examples{
good_taxon_table <- data.frame(sum.taxonomy = c("a;b;c;d;f;u", "p;q;r;s;t;u"),
site_1 = c(0,1), site_2 = c(10, 20))
scrub_taxon_paths(good_taxon_table)
}
