% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_file_validators.R
\name{validate_input_files}
\alias{validate_input_files}
\title{Verify that the input taxon_table file and the input mapping file meets specificationss
The function takes one taxon table as its input, and verfies that it meets
the expected standards.
The standards incude:
1. Column names exist.
2. One of the columns is named "sum.taxonomy"
3. The "xxx_seq_number" column, if it ever existed, is removed
4. All columns apart from sum.taxonomy should be numeric
5. All columns apart from sum.taxonomy should have corresponding row in metadata file}
\usage{
validate_input_files(taxon_table, metadata_file)
}
\arguments{
\item{taxon_table}{taxonomy table from Anacapa}

\item{metadata_file}{Qiime-style mapping}
}
\description{
Verify that the input taxon_table file and the input mapping file meets specificationss
The function takes one taxon table as its input, and verfies that it meets
the expected standards.
The standards incude:
1. Column names exist.
2. One of the columns is named "sum.taxonomy"
3. The "xxx_seq_number" column, if it ever existed, is removed
4. All columns apart from sum.taxonomy should be numeric
5. All columns apart from sum.taxonomy should have corresponding row in metadata file
}
\author{
Gaurav Kandlikar
}
