% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anacapa_utils.R
\name{vegan_otu}
\alias{vegan_otu}
\title{Takes a phyloseq object with an otu_table object and returns a vegan style community matrix.}
\usage{
vegan_otu(physeq_object)
}
\arguments{
\item{physeq_object}{phyloseq object with an otu_table object within}
}
\value{
vegan-style community matrix
}
\description{
Takes a phyloseq object with an otu_table object and returns a vegan style community matrix.
}
\author{
Gaurav Kandlikar
}
